separate http handlers into webHandlers.py module
