+++ rtCommon/projectInterface.py
@@ -22,12 +22,12 @@ from rtCommon.structDict import StructDict, recurseCreateStructDict
 from rtCommon.certsUtils import getCertPath, getKeyPath
 from rtCommon.utils import DebugLevels, writeFile, loadConfigFile
 from rtCommon.errors import StateError, RequestError, RTError
+from rtCommon.webHandlers import HttpHandler, LoginHandler, LogoutHandler, certsDir
 
-certsDir = 'certs'
+
 sslCertFile = 'rtcloud.crt'
 sslPrivateKey = 'rtcloud_private.key'
 CommonOutputDir = '/rtfmriData/'
-maxDaysLoginCookieValid = 0.5
 
 
 moduleDir = os.path.dirname(os.path.realpath(__file__))
@@ -120,10 +120,10 @@ class Web():
             # "max_buffer_size": 1024*1024*256,
         }
         Web.app = tornado.web.Application([
-            (r'/', Web.UserHttp),
-            (r'/login', Web.LoginHandler),
-            (r'/logout', Web.LogoutHandler),
-            (r'/feedback', Web.BiofeedbackHttp),  # shows image
+            (r'/', HttpHandler, dict(webObject=Web, page=Web.webIndexPage)),
+            (r'/feedback', HttpHandler, dict(webObject=Web, page=Web.webBiofeedPage)),  # shows image
+            (r'/login', LoginHandler, dict(webObject=Web)),
+            (r'/logout', LogoutHandler, dict(webObject=Web)),
             (r'/wsUser', Web.UserWebSocket),
             (r'/wsSubject', Web.BiofeedbackWebSocket),
             (r'/wsData', Web.DataWebSocket),
@@ -424,116 +424,6 @@ class Web():
                 return
         runVals.append({'x': x, 'y': y})
 
-    class UserHttp(tornado.web.RequestHandler):
-        def get_current_user(self):
-            return self.get_secure_cookie("login", max_age_days=maxDaysLoginCookieValid)
-
-        @tornado.web.authenticated
-        def get(self):
-            full_path = os.path.join(Web.htmlDir, Web.webIndexPage)
-            logging.log(DebugLevels.L6, 'Index request: pwd: {}'.format(full_path))
-            Web.httpLock.acquire()
-            try:
-                self.render(full_path)
-            finally:
-                Web.httpLock.release()
-
-    class BiofeedbackHttp(tornado.web.RequestHandler):
-        def get_current_user(self):
-            return self.get_secure_cookie("login", max_age_days=maxDaysLoginCookieValid)
-
-        @tornado.web.authenticated
-        def get(self):
-            full_path = os.path.join(Web.htmlDir, Web.webBiofeedPage)
-            logging.log(DebugLevels.L6, 'Subject feedback http request: pwd: {}'.format(full_path))
-            Web.httpLock.acquire()
-            try:
-                self.render(full_path)
-            finally:
-                Web.httpLock.release()
-
-    class LoginHandler(tornado.web.RequestHandler):
-        loginAttempts = {}
-        loginRetryDelay = 10
-
-        def get(self):
-            params = {
-                "error_msg": '',
-                "nextpage": self.get_argument("next", "/")
-            }
-            full_path = os.path.join(Web.htmlDir, Web.webLoginPage)
-            self.render(full_path,  **params)
-
-        def post(self):
-            errorReply = None
-            try:
-                login_name = self.get_argument("name")
-                login_passwd = self.get_argument("password")
-                if Web.testMode is True:
-                    if login_name == login_passwd == 'test':
-                        self.set_secure_cookie("login", login_name, expires_days=maxDaysLoginCookieValid)
-                        self.redirect(self.get_query_argument('next', '/'))
-                        return
-                passwdFilename = os.path.join(certsDir, 'passwd')
-                passwdDict = loadPasswdFile(passwdFilename)
-                if login_name in passwdDict:
-                    errorReply = self.checkRetry(login_name)
-                    if errorReply is None:
-                        hashed_passwd = passwdDict[login_name]
-                        # checkpw expects bytes array rather than string so use .encode()
-                        if bcrypt.checkpw(login_passwd.encode(), hashed_passwd.encode()) is True:
-                            # Remove failed attempts entry
-                            del Web.LoginHandler.loginAttempts[login_name]
-                            self.set_secure_cookie("login", login_name, expires_days=maxDaysLoginCookieValid)
-                            self.redirect(self.get_query_argument('next', '/'))
-                            return
-                        else:
-                            errorReply = 'Login Error: Login Incorrect'
-                else:
-                    errorReply = self.checkRetry('invalid_user')
-                    if errorReply is None:
-                        errorReply = 'Login Error: Login Incorrect'
-            except Exception as err:
-                errorReply = 'Exception: {} {}'.format(type(err), err)
-            assert errorReply is not None, "Assert: Web.LoginHandler.error not empty"
-            logging.warning('Login Failure: {}'.format(login_name))
-            params = {
-                "error_msg": errorReply,
-                "nextpage": self.get_query_argument('next', '/')
-            }
-            full_path = os.path.join(Web.htmlDir, Web.webLoginPage)
-            self.render(full_path,  **params)
-
-        def checkRetry(self, user):
-            '''Keep a dictionary with one entry per username. Any user not in the
-                passwd file will be entered as 'invalid_user'. Record login failure
-                count and timestamp for when the next retry is allowed. Reset failed
-                retry count on successful login. Return message with how many seconds
-                until next login attempt is allowed.
-            '''
-            now = time.time()
-            loginAttempts = Web.LoginHandler.loginAttempts
-            retryTime = now + Web.LoginHandler.loginRetryDelay
-            loginTry = loginAttempts.get(user)
-            if loginTry is not None:
-                failedLogins = loginTry.get('failedLogins', 0)
-                nextAllowedTime = loginTry.get('nextAllowedTime', now)
-                # print('user: {}, tries {}, nextTime {}'.format(user, failedLogins, nextAllowedTime))
-                if nextAllowedTime > now:
-                    delaySecs = loginTry['nextAllowedTime'] - now
-                    return 'Next login retry allowed in {} sec'.format(int(delaySecs))
-                loginTry['failedLogins'] = failedLogins + 1
-                loginTry['nextAllowedTime'] = retryTime
-                loginAttempts[user] = loginTry
-            else:
-                loginAttempts[user] = {'failedLogins': 1, 'nextAllowedTime': retryTime}
-            return None
-
-    class LogoutHandler(tornado.web.RequestHandler):
-        def get(self):
-            self.clear_cookie("login")
-            self.redirect("/login")
-
     class BiofeedbackWebSocket(tornado.websocket.WebSocketHandler):
         # TODO - combine these in-common setups into helper functions
         def open(self):
@@ -690,17 +580,6 @@ class Web():
                 logging.error('DataWebSocket: on_message error: {}'.format(err))
 
 
-def loadPasswdFile(filename):
-    with open(filename, 'r') as fh:
-        entries = fh.readlines()
-    passwdDict = {k: v for (k, v) in [line.strip().split(',') for line in entries]}
-    return passwdDict
-
-
-def storePasswdFile(filename, passwdDict):
-    with open(filename, 'w') as fh:
-        for k, v in passwdDict.items():
-            fh.write('{},{}\n'.format(k, v))
 
 
 def getCookieSecret(dir):
